Misc fixes for better reliability
Abbreviations that used letters from the source were having problems,
where sometimes those letters didn't appear. The problem was that the
previous key was still pressed. Now it correctly releases the previous
key before trying to type the next one.

Additionally, now _split_combination always returns raw scan codes. This
didn't simplify much of the code because we still check, but it should
be slightly faster and future code can rely on this.