Don't force canonicalized keys to be scan codes
The first step when creating a hotkey is to `canonicalize` the
combination received, which transforms key descriptions like `win` into
known scan codes like 92, then groups them in parts and steps. But there
may be *two* `win` keys on the keyboard! So when it transforms into a scan
code it has to choose one of them, and the other is ignored. Usually
this is not a problem because it chooses the left ctrl, left alt, etc,
which are the more commonly used, and why I didn't notice this problem
before. But in this case it selects the *right* `win` key (my keyboard
doesn't even have one!).

The correct behavior would be to match both keys if the user passed a
name, but only one if the user passed a scan code. Therefore the
`canonicalize` function must respect the user input type.