Ignore injected events in Windows
This should help with #48, where the alt key was appearing double. In my
computer a single click of 'alt' sent 'down, up, down', leaving the key
in a pressed state. The reporter complains about a slightly different
problem, but it's likely connected.

I noticed the second key had the INJECTED flag, so this patch ignores
injected events. Still have to watch out for unintended consequences.

Note that I'm keeping the _listener.is_replaying code that tries to
ignore replayed keys, because this fix is Windows-only.