Start major refactoring
There are two main changes:

1) Every listening function will receive a `suppress` argument, used to
choose if the callback will be blocking or not.
2) Convert names to scan codes as soon as possible. In case of multiple
scan codes for a each name, build all possible combinations.

The first change is for new features, while the second should simplify
code and make it more robuts. It's almost a rewrite though, so I'm also
taking the time to completely rewrite the keyboard tests, now with a
saner architecture.